feat: BaseProfileViewController 구현
#119
